gptbitcoin/oop/practicum/requirements.txt
yfinance==0.2.61
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
snowflake-connector-python==3.15.0